+++ CHANGELOG.md
@@ -32,6 +32,9 @@ is used.
 `cb`) arguments, rather than (`iss`, `sub`, `cb`).
 - The 4-arity form of `verify` function now invoked with (`iss`, `profile`,
 `idToken`, `cb`) arguments, rather than (`iss`, `sub`, `profile`, `cb`).
+- The 6-arity form of `verify` function now invoked with (`iss`, `profile`,
+`idToken`, `accessToken`, `refreshToken`, `cb`) arguments, rather than (`iss`,
+`sub`, `profile`, `accessToken`, `refreshToken`, `cb`).
 - `prompt` option can now take any value, rather than just defined values, in
 order to support values defined by extensions.
 - `display` option can now take any value, rather than just defined values, in